pyproject.toml
[tool.poetry]
name = "Python Projects for Learning"
version = "0.1.0"
description = "Projects for doing handson for different machine learning scenario's"
authors = ["Bharath Kumar"]

[tool.poetry.dependencies]
python = "^3.8"
dvc = "^2.10.0"
hydra-core = "^1.1.1"
pdoc3 = "^0.10.0"
notebook = "^6.4.10"
omegaconf = "^2.2.2"
interrogate = "^1.5.0"
pre-commit = "^2.19.0"
apyori = "^1.1.2"
PyYAML = "^6.0"
matplotlib = "^3.5.2"
black = "^22.3.0"
autopep8 = "^1.6.0"
pandas = "^1.4.2"


[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.17.0"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.interrogate]
ignore-init-method = true
ignore-init-module = false
ignore-magic = false
ignore-semiprivate = false
ignore-private = false
ignore-property-decorators = false
ignore-module = true
ignore-nested-functions = false
ignore-nested-classes = true
ignore-setters = false
fail-under = 95
exclude = ["setup.py", "docs", "build"]
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*"]
verbose = 0
quiet = false
whitelist-regex = []
color = true
generate-badge = "."
badge-format = "svg"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build   
)/ 
'''